fix(prefab ids): prefab window no longer creates new IDs
Editing a prefab in the editor window would spawn an ID on all UniqueID components. This was quite
annoying, as it would hard write the ID to all child prefabs as soon as it was saved. Made it so
UniqueIDs will force clear if they are being editing in the prefab window.

fix #8